refator: refatorando orderservice

+++ backend-project/src/order/order.service.ts
@@ -1,8 +1,15 @@
+import {
+  BadRequestException,
+  HttpException,
+  HttpStatus,
+  Injectable,
+  Logger,
+  Scope,
+} from '@nestjs/common';
 import { Mapper } from '@automapper/core';
 import { InjectMapper } from '@automapper/nestjs';
-import { BadRequestException, Injectable, Logger, Scope } from '@nestjs/common';
+import { CartService } from 'src/cart/cart.service';
 import { CreateOrderDto } from './dto/create-order.dto';
-import { UpdateOrderDto } from './dto/update-order.dto';
 import { Order } from './entities/order.entity';
 import { OrderStatus } from './entities/OrderStatus.enum';
 import { OrderRepository } from './order.repository';
@@ -13,31 +20,98 @@ export class OrderService {
     @InjectMapper() private mapper: Mapper,
     private readonly logger: Logger,
     private readonly repository: OrderRepository,
+    private readonly cartService: CartService,
   ) {
     this.logger = new Logger(OrderService.name);
   }
 
   async createOrder(orderModel: CreateOrderDto) {
-    return this.repository.createOrder(orderModel);
+    const orderMapper = this.mapper.map(orderModel, CreateOrderDto, Order);
+
+    orderMapper.products = await this.cartService.findCartByUser(
+      orderMapper.userId,
+    );
+    const { _id } = await this.repository.create(orderMapper);
+    let order = await this.repository.findPopulateProducts(_id);
+
+    const totalPrice = order.products.reduce((acc, product) => {
+      const price = product.price * product.quantity;
+      return acc + price;
+    }, 0);
+    await order.update({ totalPrice });
+
+    order = await this.repository.findPopulateUserProducts(_id);
+    this.logger.log('created user');
+    await order.save();
+    await this.cartService.removeAllCart(orderMapper.userId);
+
+    return order;
   }
 
-  async listOrdersByUser(user: string) {
-    return await this.repository.listOrdersByUser(user);
+  async listOrdersByUser(userId: string) {
+    this.logger.log(`looking for order with user: ${userId}`);
+    const orders = await this.repository.findPopulateUserIdProducts(userId);
+
+    if (!orders) {
+      this.logger.error(`there is no order with the user: ${userId}`);
+      throw new HttpException('No Orders Found', HttpStatus.NO_CONTENT);
+    }
+
+    this.logger.log('user found');
+    return orders;
   }
 
   async cancelledOrderByUser(orderId: string) {
-    return this.repository.cancelledOrderByUser(orderId);
+    this.logger.log(`looking for order with id: ${orderId}`);
+    const order = await this.repository.findOneUserId(orderId);
+
+    if (!order) {
+      this.logger.error(`there is no order with the id: ${orderId}`);
+      throw new BadRequestException('order not found');
+    }
+
+    this.logger.log('order found');
+    return order.updateOne({ status: OrderStatus.Cancelled });
   }
 
   async completedOrderByUser(orderId: string) {
-    return this.repository.completedOrderByUser(orderId);
+    this.logger.log(`looking for order with id: ${orderId}`);
+    const order = await this.repository.findOneUserId(orderId);
+
+    if (!order) {
+      this.logger.error(`there is no order with the id: ${orderId}`);
+      throw new BadRequestException('order not found');
+    }
+
+    this.logger.log('order found');
+    return order.updateOne({ status: OrderStatus.Completed });
   }
 
   async remove(orderId: string) {
+    this.logger.log(`looking for order with id: ${orderId}`);
+    const order = await this.repository.findById(orderId);
+
+    if (!order) {
+      this.logger.error(`there is no order with the id: ${orderId}`);
+      throw new BadRequestException('order not found');
+    }
+
+    this.logger.log('order found');
+    this.logger.log('order removed');
     return this.repository.remove(orderId);
   }
 
   async findAll() {
+    this.logger.log('Looking for all orders');
+    const order = await this.repository.findAll();
+
+    if (!order) {
+      this.logger.error('Error fetching orders');
+      throw new BadRequestException('orders not found');
+    }
+
+    this.logger.log('orders found');
+
     return await this.repository.findAll();
   }
 }
